removed useless cast

+++ plugins/org.eclipse.emf.parsley.common/src/org/eclipse/emf/parsley/builders/TableViewerColumnBuilder.java
@@ -85,7 +85,7 @@ public class TableViewerColumnBuilder {
 		for (EStructuralFeature eStructuralFeature : typeFeatures) {
 			int weight=defaultWeight;
 			if(weights.size()>i){
-				weight=(Integer)weights.get(i++);
+				weight=weights.get(i++);
 			}
 			buildTableViewerColumn(tableViewer, layout, eClass, eStructuralFeature,
 					contentProvider,weight);
